% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_correction.R
\name{build_W}
\alias{build_W}
\title{Build the full and normalised weight matrices W_n and W*_n from the
one-dimensional count matrices for each spatial dimension.}
\usage{
build_W(mat1, mat2)
}
\arguments{
\item{mat1}{Output of build_count_matrices(n1).}

\item{mat2}{Output of build_count_matrices(n2).}
}
\value{
A list with:
\item{W}{L1\emph{L2 x L1}L2 unnormalised weight matrix (I - A(x)A - B(x)B + D(x)D).}
\item{W_star}{L1\emph{L2 x L1}L2 normalised weight matrix.}
}
\description{
Build the full and normalised weight matrices W_n and W*_n from the
one-dimensional count matrices for each spatial dimension.
}
