% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{check_B_equals_JAJ}
\alias{check_B_equals_JAJ}
\title{Check the bisymmetry relation B = J A J.}
\usage{
check_B_equals_JAJ(A, B, tol = 1e-10)
}
\description{
Check the bisymmetry relation B = J A J.
}
