% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{check_irreducible}
\alias{check_irreducible}
\title{Check irreducibility of a nonnegative matrix.
A is irreducible iff (I + |A|)^{n-1} > 0.}
\usage{
check_irreducible(A, tol = 1e-10)
}
\description{
Check irreducibility of a nonnegative matrix.
A is irreducible iff (I + |A|)^{n-1} > 0.
}
