% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{check_primitive}
\alias{check_primitive}
\title{Check primitivity: irreducible + aperiodic (exists k with A^k > 0).}
\usage{
check_primitive(A, tol = 1e-10)
}
\description{
Check primitivity: irreducible + aperiodic (exists k with A^k > 0).
}
