% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_lattice.R
\name{generate_grid}
\alias{generate_grid}
\title{Create a regular rectangular lattice grid.}
\usage{
generate_grid(n1, n2 = n1)
}
\arguments{
\item{n1}{Number of rows (vertical dimension).}

\item{n2}{Number of columns (horizontal dimension). Defaults to n1.}
}
\value{
A data.frame with columns t1, t2 (integer coordinates).
}
\description{
Create a regular rectangular lattice grid.
}
