% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{get_eigenvalue_multiplicities}
\alias{get_eigenvalue_multiplicities}
\title{Cluster eigenvalues and report their multiplicities.}
\usage{
get_eigenvalue_multiplicities(eigenvalues, tol = 1e-10)
}
\description{
Cluster eigenvalues and report their multiplicities.
}
