% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_checks.R
\name{get_spectral_summary}
\alias{get_spectral_summary}
\title{Comprehensive spectral summary of a square matrix.}
\usage{
get_spectral_summary(A, tol = 1e-10)
}
\description{
Comprehensive spectral summary of a square matrix.
}
