% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Jmat}
\alias{Jmat}
\title{Build the \( n \times n \) reversal matrix \( \mathbf{J} \) (anti-identity).}
\usage{
Jmat(n)
}
\arguments{
\item{n}{Integer. Dimension of the matrix.}
}
\value{
An \( n \times n \) numeric matrix representing the reversal operator.
}
\description{
The matrix has ones on the anti-diagonal and zeros elsewhere.
Multiplication \( \mathbf{J} \\%*\\% \underline{x} \) reverses the entries of a vector \( \underline{x} \).
}
