% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Convert a matrix with dimnames "h1=..." / "h2=..." to a long data frame.}
\usage{
matrix_to_df(mat, name_val = "val")
}
\arguments{
\item{mat}{Matrix with named rows and columns.}

\item{name_val}{Name for the value column (default "val").}
}
\description{
Convert a matrix with dimnames "h1=..." / "h2=..." to a long data frame.
}
