% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse_ratio}
\alias{mse_ratio}
\title{Compute the ratio tr(MSE_corrected) / tr(MSE_standard).
Values < 1 indicate the bias-corrected estimator is better.}
\usage{
mse_ratio(
  n1,
  n2,
  m,
  beta,
  lambdas,
  alphas,
  sigma,
  mu = 0,
  compact = TRUE,
  Nh_normalize = TRUE
)
}
\description{
Compute the ratio tr(MSE_corrected) / tr(MSE_standard).
Values < 1 indicate the bias-corrected estimator is better.
}
