% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{scalar_mse}
\alias{scalar_mse}
\title{Extract the scalar (trace) MSE from an MSE matrix.
Equivalent to sum of diagonal = sum of individual MSE(Chat(h)) over all h.}
\usage{
scalar_mse(MSE_matrix)
}
\description{
Extract the scalar (trace) MSE from an MSE matrix.
Equivalent to sum of diagonal = sum of individual MSE(Chat(h)) over all h.
}
