% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{verify_compact_vs_16term}
\alias{verify_compact_vs_16term}
\title{Check that the compact (mu-free) and 16-term formulas produce the same
Omega matrix, for a given set of parameters. Returns the max absolute
difference.}
\usage{
verify_compact_vs_16term(n1, n2, m, mu, beta, lambdas, alphas, sigma)
}
\description{
Check that the compact (mu-free) and 16-term formulas produce the same
Omega matrix, for a given set of parameters. Returns the max absolute
difference.
}
